requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
alembic==1.16.4
    # via -r requirements.in
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via starlette
asyncpg==0.30.0
    # via -r requirements.in
bcrypt==4.0.1
    # via
    #   -r requirements.in
    #   passlib
cffi==1.17.1
    # via cryptography
click==8.2.1
    # via uvicorn
cryptography==45.0.5
    # via python-jose
dnspython==2.7.0
    # via email-validator
ecdsa==0.19.1
    # via python-jose
email-validator==2.2.0
    # via -r requirements.in
fastapi==0.116.1
    # via -r requirements.in
greenlet==3.2.3
    # via sqlalchemy
gunicorn==23.0.0
    # via -r requirements.in
h11==0.16.0
    # via uvicorn
idna==3.10
    # via
    #   anyio
    #   email-validator
mako==1.3.10
    # via alembic
markupsafe==3.0.2
    # via mako
packaging==25.0
    # via gunicorn
passlib[bcrypt]==1.7.4
    # via -r requirements.in
pyasn1==0.6.1
    # via
    #   python-jose
    #   rsa
pycparser==2.22
    # via cffi
pydantic==2.11.7
    # via
    #   fastapi
    #   pydantic-settings
    #   sqlmodel
pydantic-core==2.33.2
    # via pydantic
pydantic-settings==2.10.1
    # via -r requirements.in
python-dotenv==1.1.1
    # via pydantic-settings
python-jose[cryptography]==3.5.0
    # via -r requirements.in
python-multipart==0.0.20
    # via -r requirements.in
rsa==4.9.1
    # via python-jose
six==1.17.0
    # via ecdsa
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.42
    # via
    #   alembic
    #   sqlmodel
sqlmodel==0.0.24
    # via -r requirements.in
starlette==0.47.2
    # via fastapi
typing-extensions==4.14.1
    # via
    #   alembic
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   starlette
    #   typing-inspection
typing-inspection==0.4.1
    # via
    #   pydantic
    #   pydantic-settings
uvicorn==0.35.0
    # via -r requirements.in
